+++ packages/react/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.203.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react@0.202.0...@carbon-labs/react@0.203.0) (2026-08-20)
+
+
+### Features
+
+* **animated-header:** animated header carousel auto pagination ([#1347](https://github.com/carbon-design-system/carbon-labs/issues/1347)) ([bb5e981](https://github.com/carbon-design-system/carbon-labs/commit/bb5e98154ebb9c8c727770f06569b58cfc36e8ae))
+
+
+
+
+
+# [0.202.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react@0.201.0...@carbon-labs/react@0.202.0) (2026-08-19)
+
+
+### Features
+
+* **animated-header:** Update header to have page carousel ([#1300](https://github.com/carbon-design-system/carbon-labs/issues/1300)) ([3d3d106](https://github.com/carbon-design-system/carbon-labs/commit/3d3d106e249bbd676843ab2bd2085651ffa363c1))
+
+
+
+
+
 # [0.201.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react@0.200.0...@carbon-labs/react@0.201.0) (2026-08-13)
 
 **Note:** Version bump only for package @carbon-labs/react